src/main/java/net/ancronik/cookbook/backend/web/controller/RecipeController.java
package net.ancronik.cookbook.backend.web.controller;

import net.ancronik.cookbook.backend.application.exceptions.EmptyDataException;
import net.ancronik.cookbook.backend.domain.service.RecipeService;
import net.ancronik.cookbook.backend.web.dto.RecipeDto;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

/**
 * Controller handling all operations for the {@link net.ancronik.cookbook.backend.data.model.Recipe}.
 *
 * @author Nikola Presecki
 */
@RestController
@RequestMapping("/api/v1/recipes")
public class RecipeController {

    private final RecipeService recipeService;

    @Autowired
    public RecipeController(RecipeService recipeService) {
        this.recipeService = recipeService;
    }

    @GetMapping(value = {"", "/"})
    public Page<RecipeDto> getAllRecipes(Pageable pageable) {
        Page<RecipeDto>  page = recipeService.getAllRecipes(pageable.getPageNumber(), pageable.getPageSize(), pageable.getSort());

        if(null == page){
            throw new EmptyDataException("Empty page returned by service");
        }

        return page;
    }
}
